Second Commit
Update on System

+++ Utilitario.java
@@ -1,8 +1,8 @@
 package Classes;
 
 public class Utilitario{
-	public static double custoBasico = 100.0;
-	public static double custoDependente = 40.0;
-	public static double custoAcademia = 90.0;
-	public static double custoNatacao = 60.0;
+	public static float custoBasico = 100;
+	public static float custoDependente = 40;
+	public static float custoAcademia = 90;
+	public static float custoNatacao = 60;
 }
